Former-commit-id: eb25f85ffd209c91872cb28bf7bdb2c17beef471

+++ Back-End/ssachedule/src/main/java/com/ssljjong/ssachedule/dto/TrackDto.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class TrackDto {
 
-    private Long id;
+    private Integer id;
     private String name;
     private int gi;
 
-    public TrackDto(Long id, String name, int gi) {
+    public TrackDto(Integer id, String name, int gi) {
         this.id = id;
         this.name = name;
         this.gi = gi;
